Refactoring and re-naming
