+++ examples/readyaml_buildmesh.py
@@ -3,7 +3,7 @@ import numpy as np
 import pickle
 from pprint import pprint
 
-from pynumad.shell.shell import getShellMesh
+from pynumad.shell.shell import get_shell_mesh
 
 blade = pynu.Blade()
 fileName = 'example_data/myBlade.yaml'
@@ -12,7 +12,7 @@ blade.read_yaml(fileName)
 elementSize = 0.2
 adhes = 1
 
-meshData = getShellMesh(blade, adhes, elementSize)
+meshData = get_shell_mesh(blade, adhes, elementSize)
 
 # Print first 10 nodes coordinates
 pprint(meshData['nodes'][:10,:])
