+++ examples/write_abacus_solid_model.py
@@ -2,7 +2,7 @@ import pynumad as pynu
 import numpy as np
 import os
 
-from pynumad.shell.shell import getSolidMesh
+from pynumad.shell.shell import get_solid_mesh
 
 ## Read blade data from yaml file
 blade = pynu.objects.Blade.Blade()
@@ -23,7 +23,7 @@ elementSize = 0.2
 
 ## Specify the elements per primary layer and generate mesh
 layNumEls = [1,1,1]
-bladeMesh = getSolidMesh(blade,layNumEls,elementSize)
+bladeMesh = get_solid_mesh(blade,layNumEls,elementSize)
 
 ## Write mesh to yaml
 meshFile = 'solidMeshData.yaml'
