+++ src/pynumad/objects/Material.py
@@ -1,10 +1,3 @@
-########################################################################
-#                    Part of the SNL NuMAD Toolbox                     #
-#  Developed by Sandia National Laboratories Wind Energy Technologies  #
-#              See license.txt for disclaimer information              #
-########################################################################
-
-
 class Material:
     """MaterialDef:  A class for blade materials.
