+++ src/pynumad/objects/Subobjects.py
@@ -1,10 +1,3 @@
-########################################################################
-#                    Part of the SNL NuMAD Toolbox                     #
-#  Developed by Sandia National Laboratories Wind Energy Technologies  #
-#              See license.txt for disclaimer information              #
-########################################################################
-
-
 class MatDBentry:
     """A simple class to organize the attributes of a material"""
