+++ src/pynumad/shell/segment2d.py
